Convert to pixel

+++ src/uet/oop/bomberman/entities/Balloon.java
@@ -3,7 +3,7 @@ package uet.oop.bomberman.entities;
 import javafx.scene.image.Image;
 
 public class Balloon extends Entity {
-    public Balloon(double x, double y, Image img) {
+    public Balloon(int x, int y, Image img) {
         super(x, y, img);
     }
