+++ src/uet/oop/bomberman/entities/Oneal.java
@@ -4,7 +4,7 @@ import javafx.scene.image.Image;
 import uet.oop.bomberman.BombermanGame;
 
 public class Oneal extends Entity {
-    public Oneal(double x, double y, Image img) {
+    public Oneal(int x, int y, Image img) {
         super(x, y, img);
     }
     @Override
