+++ src/uet/oop/bomberman/entities/Portal.java
@@ -3,7 +3,7 @@ package uet.oop.bomberman.entities;
 import javafx.scene.image.Image;
 
 public class Portal extends Entity {
-    public Portal(double x, double y, Image img) {
+    public Portal(int x, int y, Image img) {
         super(x, y, img);
     }
